chore: check whether react-query wrapper has accessToken present (#882)
Because

- Better debug

This commit

- check whether react-query wrapper has accessToken present

+++ packages/toolkit/src/lib/react-query-service/mgmt/useDeleteApiToken.ts
@@ -12,6 +12,10 @@ export const useDeleteApiToken = () => {
       tokenName: string;
       accessToken: Nullable<string>;
     }) => {
+      if (!accessToken) {
+        return Promise.reject(new Error("accessToken not provided"));
+      }
+
       await deleteApiTokenMutation({ tokenName, accessToken });
       return Promise.resolve(tokenName);
     },
